requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Cython==0.29.26
ghp-import==2.0.2
idna==3.3
importlib-metadata==4.8.3
install==1.3.5
Jinja2==3.0.3
jsonschema==4.2.1
Markdown==3.3.6
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.3
mkdocs-autorefs==0.3.0
mkdocs-include-markdown-plugin==3.2.3
mkdocs-material==8.1.2
mkdocs-material-extensions==1.0.3
mkdocstrings==0.16.2
moves==0.1
numpy==1.20.0
packaging==21.3
pandas==1.3.5
Pygments==2.10.0
pymdown-extensions==9.1
pyparsing==3.0.6
pyproject-toml==0.0.11
pyrsistent==0.18.0
python-dateutil==2.8.2
pytkdocs==0.12.0
pytz==2021.3
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.26.0
six==1.16.0
toml==0.10.2
trafaret==2.1.0
urllib3==1.26.7
watchdog==2.1.6
zipp==3.6.0
pycryptodome==3.12.0
